Fix issue #7
Fixes the issue of apostrophes in an article titles causing the title to be truncated in search results. 
Resolves #7 (https://github.com/joomla-extensions/search/issues/7)
Resolves joomla/joomla-cms#34724 (https://github.com/joomla/joomla-cms/issues/34724)
Work around is to not run titles through search preparation which truncates text around the search term. Titles are returned in full instead - which is the expected behavior for search results (a user wants to see the title of an article not an extract of the title).